Move results/vacancy functions from results route into module
