+++ routes/results.js
@@ -1,23 +1,21 @@
 const express = require('express');
 const router = express.Router();
 const url = require('url');
-const fetch = require('node-fetch');
 
-const pager = require('../lib/modules/pagination');
+const Pager = require('../lib/modules/pagination');
+const { fetchVacancyList, formatResultData } = require('../lib/modules/vacancy');
 const UrlUtils = require('../lib/modules/url');
-
-const VIEW_PATH = 'pages/results';
 
 /* GET results page. */
 router.get('/', function(req, res) {
-  const { location, keyword, page = 0 } = url.parse(req.url, true).query;
+  const filters = url.parse(req.url, true).query;
   const queryString = url.parse(req.url).query;
 
   try {
-    fetchVacancyList(`${process.env.API_URL}:${process.env.API_PORT}/vacancy/search/location/${location || null}/keyword/${keyword || null}?page=${page - 1}`).then(data => {
+    fetchVacancyList(filters).then(data => {
       
       const url = `/results?${ UrlUtils.removeUrlParameter(queryString, 'page') }`;
-      const pagerOptions = pager(
+      const pagerOptions = Pager(
           data.totalPages, 
           data.first, 
           data.last, 
@@ -28,14 +26,14 @@ router.get('/', function(req, res) {
           data.number
         );
       
-      res.render(VIEW_PATH, {
+      res.render('pages/results', {
         i18n: {
           ...req.translations,
           title: req.translations.results.page.title,
           resultsTotal: data.totalElements === 1 ? req.translations.results.page.totalJobsFoundSingular : req.translations.results.page.totalJobsFoundPlural
         },
         results: formatResultData(data.content),
-        filters: { location, keyword },
+        filters,
         returnUrl: queryString,
         pager: pagerOptions
       })
@@ -46,35 +44,5 @@ router.get('/', function(req, res) {
   }
   
 });
-
-async function fetchVacancyList(url) {
-  let response = await fetch(url);
-  return await response.json();
-}
-
-function formatSalaryNumber(num) {
-  return num.toLocaleString();
-}
-
-function formatResultData(results = []) {
-  return results.map(result => {
-    const salary = formatSalaryOutput(result.salaryMin, result.salaryMax);
-    return { ...result, salary }
-  });
-}
-
-function formatSalaryOutput(min, max) {
-  if(min !== 0 && max !== 0) {
-    return `£${formatSalaryNumber(min)} - £${formatSalaryNumber(max)}`;
-  }
-
-  if(max === 0) {
-    return `£${formatSalaryNumber(min)}`;
-  }
-
-  if(min === 0) {
-    return `£${formatSalaryNumber(max)}`;
-  }
-}
 
 module.exports = router;
